app/requirements.txt
requests
psycopg2
django
djangorestframework
django-cors-headers
